+++ src/ResourceManager/Automation/Commands.Automation/Cmdlet/RemoveAzureAutomationConnection.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Azure.Commands.Automation.Cmdlet
     /// Removes a Connection for automation.
     /// </summary>
     [Cmdlet(VerbsCommon.Remove, "AzureRmAutomationConnection", SupportsShouldProcess = true,
-        DefaultParameterSetName = AutomationCmdletParameterSets.ByName)]
+        DefaultParameterSetName = AutomationCmdletParameterSets.ByName), OutputType(typeof(void))]
     public class RemoveAzureAutomationConnection : AzureAutomationBaseCmdlet
     {
         /// <summary>