+++ src/StackAdmin/Azs.KeyVault.Admin/Tests/src/Common.ps1
@@ -1,23 +1,13 @@
 ﻿
-$global:Location = "local"
-$global:TenantVMName = "502828aa-de3a-4ba9-a66c-5ae6d49589d7"
-$global:Provider = "Microsoft.KeyVault.Admin"
+$global:Location = 'local'
 
-if(-not $RunRaw) {
+if(-not $global:RunRaw) {
 	$scriptBlock = {
 		Get-MockClient -ClassName 'KeyVaultAdminClient' -TestName $global:TestName -Verbose
 	}
-	Mock New-ServiceClient $scriptBlock -ModuleName "Azs.KeyVault.Admin"
+	Mock New-ServiceClient $scriptBlock -ModuleName $global:ModuleName
 }
 
-function Repeat{
-	param(
-		[int]$Times,
-		[ScriptBLock]$Script
-	)
-
-	while($Times -gt 0) {
-		Invoke-Command -ScriptBlock $Script
-		$Times = $Times - 1
-	}
+if (Test-Path "$PSScriptRoot\Override.ps1") {
+    . $PSScriptRoot\Override.ps1
 }