+++ src/ResourceManager/EventHub/Commands.EventHub/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Updated all help files to include full parameter types and correct input/output types.
+
+## Version 0.6.6
 * Added a readonly property 'PendingReplicationOperationsCount' to PSEventHubDRConfigurationAttributes class, which gives the pending replication operations count while replication is in progress
 
 ## Version 0.6.5