+++ src/ResourceManager/Profile/Commands.Profile/ChangeLog.md
@@ -18,6 +18,11 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Updated Common.Strategy library to be able to validate that the current config for a resource is compatible with the target resource. Default is always true, individual resources and overridet the default.
+* Updated all help files to include full parameter types and correct input/output types.
+* Added ps1xml types to Common.Storage
+
+## Version 5.3.3
 * Updated help for `Get-AzureRmContextAutosaveSetting`
 
 ## Version 5.3.2