+++ src/ResourceManager/Resources/Commands.Resources/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Updated help files to include full parameter types and correct input/output types.
+* Fix piping scenario with `Set-AzureRmResource`
+
+## Version 6.2.1
 * Update Roleassignment and roledefinition cmdlets:
     - Remove extra roledefinition calls done as part of paging.
 * Fix Get-AzureRmRoleAssignment cmdlet